refactor: move schemes to new file
